+++ DatBot.Interface/src/protocol/network/util/types/Int64.java
@@ -6,16 +6,16 @@ public class Int64 extends Binary64{
 		
 	}
 	
-	public Int64(long n, int high){
+	public Int64(int n, int high){
 		super(n,high);
 	}
 	
     public static Int64 fromNumber(long n)
     {
-        return new Int64((long) n, (int) Math.floor(n / 4.294967296E9));
+        return new Int64((int) n, (int) Math.floor(n / 4.294967296E9));
     }
     
-    public long getHigh(){
+    public int getHigh(){
     	return this.high;
     }
